.history/requirements_20250730212108.txt
streamlit
chromadb
langchain
ollama
pypdf
sentence-transformers
